New add-image page [#5]
Note: in order to style the file input, there's a funny hack in place.
The file input sits in front of a button (actually a <label>), but has
opacity 0; its width and position need to be aligned just right with the
"select" button.

The nav-tabs' border rule needs extra specificity to override
bootstrap's defaults, presumably because CSS was designed by malevolent
spirits.

When the controller receives a file, it calculates a filename and tells
the page JS about it. From that point on, that has to be the filename;
recalculating it after reorienting isn't ok. So the filename is now
frozen on ImageTruck initialization.