Remove nested forms in the edit-after-add pane [#45]
tagLink returns an element containing a form, so appending it to a form
results in a nested-forms situation, where behavior is unpredictable.

Instead, the title and description areas get individual forms. That
means that when saving attributes, we have to go to the forms' parent to
find the individual inputs.

The "enter" key-listener was a bad fix for a problem introduced by the
nested forms, so it can go away now.