Only click "add" when tabbing from new-tag input [#45]
The on-tab handler tries to avoid complexity by just finding the "add
another tag" link and clicking it, rather than knowing anything about
the internals of the new-tag-input logic. Unfortunately it was also
finding all the remove-tag links...

No tests for this change, because I can't get Selenium to reproduce the
problem. :(