Notify the browser when images have been uploaded [#40]
The new system goes like this:
* AJAX request to upload an image
* Server starts a task to process the image
* Server responds with the id of the celery task
* JS opens a WebSocket connection to the server
* JS sends a WS message asking about the celery task id
* Server receives message and redis-subscribes for messages regarding
    that task id.
* Task finishes, at some point. The worker redis-publishes that it's
    done (once per resize & once for the original).
* Server receives redis message and sends WS to the client.
* Client tries to load the image.

celery_broker_url is now redis_url. I'm not sure if the code would've
worked with a non-redis broker before, but now it's non-optional for
sure.

All the logic for keeping celery jobs in sync with the contents of the
database is now in db_redis_coordination.py. Tasks get a transaction_id,
which corresponds to a row in the database. If that transaction_id is
present, whatever process prompted the task must have finished. If not,
the task will need to retry.