Remove command-line operation [#56]
There aren't users for the command-line client, and it really isn't the
area I want to focus on right now. Having the need for argument config
will complicate the changes I have in mind so I'm cutting out the
feature.

This also means dumping HMAC auth and the api key/host settings, since
the command line client was the only thing using any of them.